requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0
requests==2.31.0
python-engineio==4.7.1
websocket-client==1.6.4